feat(settings): extract AI sub-screen with wardrobe index tracking
Pull all AI, image enrichment, and embedding index items out of the flat
Settings screen into a dedicated AI & Image Tools sub-screen:

- AiSettingsScreen.kt (new): Assistant toggle/vibe, Provider picker +
  credential fields, Wardrobe Index progress + Rebuild trigger, and
  Image Enrichment (background removal, image descriptions) — all with
  previews
- SettingsScreen.kt: now 670 lines (was 1731); AI section replaced by a
  single navigable "AI & Image Tools" row; Wardrobe section keeps only
  Laundry Day
- EmbeddingScheduler: add runNow() for user-triggered immediate runs via
  a separate IMMEDIATE_NAME unique work (no charging/idle constraints)
- SettingsViewModel: inject EmbeddingScheduler + EmbeddingIndex; expose
  embeddingWorkInfo, embeddingIndexSize, triggerEmbeddingRebuild()
- SettingsNavigation: add AiSettingsRoute + aiSettingsScreen() extension
- ClosetNavGraph: register AI settings screen
