+++ app-android/features/settings/src/main/kotlin/com/closet/features/settings/SettingsNavigation.kt
@@ -10,6 +10,10 @@ import kotlinx.serialization.Serializable
 @Serializable
 object SettingsRoute
 
+/** Route for the AI & Image Tools sub-screen. */
+@Serializable
+object AiSettingsRoute
+
 /**
  * Navigates to [SettingsRoute], optionally applying [navOptions] for back-stack
  * management (e.g. `singleTop`, `popUpTo`).
@@ -18,20 +22,41 @@ fun NavController.navigateToSettings(navOptions: NavOptions? = null) {
     navigate(SettingsRoute, navOptions)
 }
 
+/** Navigates to the [AiSettingsRoute] sub-screen. */
+fun NavController.navigateToAiSettings() {
+    navigate(AiSettingsRoute)
+}
+
 /**
  * Registers the [SettingsRoute] composable destination in the [NavGraphBuilder].
  *
  * @param onNavigateUp         Called when the user taps the back arrow.
  * @param onNavigateToBulkWash Called when the user taps the Laundry Day row.
+ * @param onNavigateToAiSettings Called when the user taps the AI & Image Tools row.
  */
 fun NavGraphBuilder.settingsScreen(
     onNavigateUp: () -> Unit,
     onNavigateToBulkWash: () -> Unit = {},
+    onNavigateToAiSettings: () -> Unit = {},
 ) {
     composable<SettingsRoute> {
         SettingsScreen(
             onNavigateUp = onNavigateUp,
             onNavigateToBulkWash = onNavigateToBulkWash,
+            onNavigateToAiSettings = onNavigateToAiSettings,
         )
+    }
+}
+
+/**
+ * Registers the [AiSettingsRoute] composable destination in the [NavGraphBuilder].
+ *
+ * @param onNavigateUp Called when the user taps the back arrow.
+ */
+fun NavGraphBuilder.aiSettingsScreen(
+    onNavigateUp: () -> Unit,
+) {
+    composable<AiSettingsRoute> {
+        AiSettingsScreen(onNavigateUp = onNavigateUp)
     }
 }
